registerLifecycle(lifecycleOwner: LifecycleOwner) method added & Java sample converted into a Fragment example
