+++ gradle.properties
@@ -24,7 +24,7 @@ kotlin.code.style=official
 # ----------------
 GROUP=org.imaginativeworld.whynotimagecarousel
 POM_ARTIFACT_ID=whynotimagecarousel
-VERSION_NAME=2.0.1
+VERSION_NAME=2.0.2
 # ----------------
 POM_NAME=Why Not! Image Carousel!
 POM_DESCRIPTION=An easy, super simple and customizable image carousel view for Android.
