+++ sample/build.gradle
@@ -65,6 +65,9 @@ dependencies {
     implementation 'com.github.bumptech.glide:glide:4.12.0'
     kapt 'com.github.bumptech.glide:compiler:4.12.0'
 
+    // Kotlin
+    implementation "androidx.fragment:fragment-ktx:1.3.4"
+
     // debugImplementation because LeakCanary should only run in debug builds.
     debugImplementation 'com.squareup.leakcanary:leakcanary-android:2.7'
 }
