+++ sample/src/main/res/layout/activity_java.xml
@@ -1,110 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="#eee"
     tools:context=".JavaActivity">
 
-    <TextView
-        android:id="@+id/tv_caption"
+    <androidx.fragment.app.FragmentContainerView
+        android:id="@+id/fragment_container_view"
+        android:name="org.imaginativeworld.whynotimagecarousel.sample.SampleFragment"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:padding="32dp"
-        android:text="@string/library_name"
-        android:textAlignment="center"
-        android:textColor="#555"
-        android:textSize="24sp"
-        android:textStyle="bold"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <com.google.android.material.card.MaterialCardView
-        android:id="@+id/carousel_container"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_marginStart="16dp"
-        android:layout_marginEnd="16dp"
-        app:cardCornerRadius="16dp"
-        app:layout_constraintDimensionRatio="3:2"
-        app:layout_constraintTop_toBottomOf="@+id/tv_caption">
-
-        <org.imaginativeworld.whynotimagecarousel.ImageCarousel
-            android:id="@+id/carousel"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent" />
-
-    </com.google.android.material.card.MaterialCardView>
-
-    <me.relex.circleindicator.CircleIndicator2
-        android:id="@+id/custom_indicator"
-        android:layout_width="match_parent"
-        android:layout_height="32dp"
-        app:ci_drawable="@drawable/dot"
-        app:ci_drawable_unselected="@drawable/dot"
-        app:layout_constraintTop_toBottomOf="@+id/carousel_container" />
-
-    <com.google.android.material.card.MaterialCardView
-        android:layout_width="match_parent"
-        android:layout_height="64dp"
-        android:layout_marginStart="16dp"
-        android:layout_marginEnd="16dp"
-        app:cardCornerRadius="32dp"
-        app:layout_constraintTop_toBottomOf="@+id/custom_indicator">
-
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent">
-
-            <com.google.android.material.button.MaterialButton
-                android:id="@+id/btn_goto_previous"
-                style="@style/Widget.App.Button.OutlinedButton.IconOnly"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="16dp"
-                android:layout_marginTop="8dp"
-                android:layout_marginBottom="8dp"
-                app:backgroundTint="@color/flat_green_2"
-                app:icon="@drawable/ic_arrow_back"
-                app:iconTint="@color/white"
-                app:layout_constraintBottom_toBottomOf="@+id/fab_play"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="@+id/fab_play" />
-
-            <com.google.android.material.button.MaterialButton
-                android:id="@+id/btn_goto_next"
-                style="@style/Widget.App.Button.OutlinedButton.IconOnly"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="8dp"
-                android:layout_marginEnd="16dp"
-                android:layout_marginBottom="8dp"
-                app:backgroundTint="@color/flat_green_2"
-                app:icon="@drawable/ic_arrow_forward"
-                app:iconTint="@color/white"
-                app:layout_constraintBottom_toBottomOf="@+id/fab_play"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintTop_toTopOf="@+id/fab_play" />
-
-            <com.google.android.material.floatingactionbutton.ExtendedFloatingActionButton
-                android:id="@+id/fab_play"
-                style="@style/Widget.MaterialComponents.ExtendedFloatingActionButton"
-                android:layout_width="wrap_content"
-                android:layout_height="32dp"
-                android:layout_marginTop="8dp"
-                android:layout_marginBottom="8dp"
-                android:gravity="center"
-                android:padding="0dp"
-                android:text="Start"
-                android:textColor="@color/white"
-                app:backgroundTint="@color/flat_red_1"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent" />
-
-        </androidx.constraintlayout.widget.ConstraintLayout>
-
-    </com.google.android.material.card.MaterialCardView>
+        android:layout_height="match_parent"
+        tools:layout="@layout/fragment_sample" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
