+++ sample/src/main/res/values/styles.xml
@@ -20,4 +20,9 @@
         <item name="iconGravity">textStart</item>
     </style>
 
+    <style name="RoundedShape">
+        <item name="cornerFamily">rounded</item>
+        <item name="cornerSize">16dp</item>
+    </style>
+
 </resources>
